+++ client/components/cardsSection.js
@@ -4,6 +4,7 @@ import Container from "./container";
 import { Swiper, SwiperSlide } from "swiper/react";
 import "swiper/css/navigation";
 import 'swiper/css';
+import Link from 'next/link';
 
 export default function cardsSection() {
     
@@ -13,8 +14,19 @@ export default function cardsSection() {
           <div className="grid lg:grid-cols-3 gap-6">
               {cardList.map((card) => (
                 
-            <div className ="shadow-lg rounded-lg">
-              <img className = "rounded-t-2xl"  src={card.img} alt="" />
+                <div className="shadow-lg rounded-lg" key={card.id}>
+                <Link href={`/home/${card.id}`}>
+                  
+                    <Image
+                      src={card.img}
+                      width={816}
+                      height={817}
+                      className="rounded-t-2xl"
+                      alt=""
+                      loading="eager"
+                    />
+                  
+                </Link>
               <div className='p-5'>
                 <h3 className='text-3xl font-bold text-slate-700 mb-3 dark:text-white'>{card.title}</h3>
                 <p className='text-lg font-normal text-gray-600 dark:text-white'>{card.text}</p>
